Article consistency
I kindly request the update of the current article as on "Is delegated access supported?" we state that it isn't but below on "Can I open encrypted messages sent to another user's mailbox with Fullaccess?" you state yes. This is inacurate and inconsistent. You can access encrypted content with full access as long as you use OWA. The scenario itself is not supported as on the clients, various additional verifications are performed that aren't performed on the shared mailbox scenario (e.g. licensing).

Please consult all 3 topics. I was the one reportingthe automapping and asked this article to be changed but it seems it was misunderstood. the text "Users can open encrypted messages as long as they are given direct access and automapping is turned ON. Access is not allowed if the access is granted via an email-enabled security group." which is currently under the topic "Can I open encrypted messages sent to another user's mailbox with Fullaccess?" was and is to be added to the shared mailbox scenario, not here.

This generates confusion for customers which not only mention that the info is inconsistent as they complain the article is in fact confusing and misleading.

Please update the same